+++ examples/ffva/bsp_config/XK_VOICE_L71/platform/platform_start.c
@@ -1,4 +1,4 @@
-// Copyright 2022-2023 XMOS LIMITED.
+// Copyright 2022-2024 XMOS LIMITED.
 // This Software is subject to the terms of the XMOS Public Licence: Version 1.
 
 /* System headers */
@@ -141,6 +141,13 @@ static void usb_start(void)
 #endif
 }
 
+static void uart_start(void)
+{
+#if ON_TILE(UART_TILE_NO)
+    rtos_uart_tx_start(uart_tx_ctx);
+#endif
+}
+
 void platform_start(void)
 {
     rtos_intertile_start(intertile_ctx);
@@ -155,4 +162,5 @@ void platform_start(void)
     mics_start();
     i2s_start();
     usb_start();
+    uart_start();
 }