+++ examples/ffva/ffva.cmake
@@ -1,11 +1,12 @@
 #**********************
 # Gather Sources
 #**********************
-file(GLOB_RECURSE APP_SOURCES ${CMAKE_CURRENT_LIST_DIR}/src/*.c )
+
+file(GLOB_RECURSE APP_SOURCES   ${CMAKE_CURRENT_LIST_DIR}/src/*.c)
+
 set(APP_INCLUDES
     ${CMAKE_CURRENT_LIST_DIR}/src
     ${CMAKE_CURRENT_LIST_DIR}/src/usb
-    ${CMAKE_CURRENT_LIST_DIR}/src/ww_model_runner
 )
 
 include(${CMAKE_CURRENT_LIST_DIR}/bsp_config/bsp_config.cmake)
@@ -36,6 +37,7 @@ set(APP_COMPILE_DEFINITIONS
 set(APP_LINK_OPTIONS
     -lquadspi
     -report
+    -lotp3
     ${CMAKE_CURRENT_LIST_DIR}/src/config.xscope
 )
 
@@ -81,6 +83,7 @@ endif()
 # XMOS Example Design Targets
 #**********************
 include(${CMAKE_CURRENT_LIST_DIR}/ffva_int.cmake)
+include(${CMAKE_CURRENT_LIST_DIR}/ffva_int_cyberon.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/ffva_ua.cmake)
 
 #**********************