+++ examples/ffva/src/app_conf.h
@@ -1,4 +1,4 @@
-// Copyright 2022-2023 XMOS LIMITED.
+// Copyright 2022-2024 XMOS LIMITED.
 // This Software is subject to the terms of the XMOS Public Licence: Version 1.
 
 #ifndef APP_CONF_H_
@@ -15,10 +15,15 @@
 #define appconfAUDIOPIPELINE_PORT      7
 #define appconfI2S_OUTPUT_SLAVE_PORT   8
 
+#ifndef appconfINTENT_ENGINE_READY_SYNC_PORT
+#define appconfINTENT_ENGINE_READY_SYNC_PORT      18
+#endif /* appconfINTENT_ENGINE_READY_SYNC_PORT */
+
 /* Application tile specifiers */
 #include "platform/driver_instances.h"
-#define FS_TILE_NO              FLASH_TILE_NO
-#define AUDIO_PIPELINE_TILE_NO  MICARRAY_TILE_NO
+#define ASR_TILE_NO                     FLASH_TILE_NO
+#define FS_TILE_NO                      FLASH_TILE_NO
+#define AUDIO_PIPELINE_OUTPUT_TILE_NO   FLASH_TILE_NO
 
 /* Audio Pipeline Configuration */
 #define appconfAUDIO_CLOCK_FREQUENCY            MIC_ARRAY_CONFIG_MCLK_FREQ
@@ -27,6 +32,61 @@
 #define appconfAUDIO_PIPELINE_CHANNELS          MIC_ARRAY_CONFIG_MIC_COUNT
 /* If in channel sample format, appconfAUDIO_PIPELINE_FRAME_ADVANCE == MIC_ARRAY_CONFIG_SAMPLES_PER_FRAME*/
 #define appconfAUDIO_PIPELINE_FRAME_ADVANCE     MIC_ARRAY_CONFIG_SAMPLES_PER_FRAME
+
+/* Enable audio response output */
+#ifndef appconfAUDIO_PLAYBACK_ENABLED
+#define appconfAUDIO_PLAYBACK_ENABLED           1
+#endif
+
+/* Intent Engine Configuration */
+#define appconfINTENT_FRAME_BUFFER_MULT      (8*2)       /* total buffer size is this value * MIC_ARRAY_CONFIG_SAMPLES_PER_FRAME */
+#define appconfINTENT_SAMPLE_BLOCK_LENGTH    240
+
+/* Maximum delay between a wake up phrase and command phrase */
+#ifndef appconfINTENT_RESET_DELAY_MS
+#if appconfAUDIO_PLAYBACK_ENABLED
+#define appconfINTENT_RESET_DELAY_MS         5000
+#else
+#define appconfINTENT_RESET_DELAY_MS         4000
+#endif
+#endif
+
+/* Output raw inferences, if set to 0, a state machine requires a wake up phrase
+ * before a command phrase */
+#ifndef appconfINTENT_RAW_OUTPUT
+#define appconfINTENT_RAW_OUTPUT   0
+#endif
+
+/* Maximum number of detected intents to hold */
+#ifndef appconfINTENT_QUEUE_LEN
+#define appconfINTENT_QUEUE_LEN     10
+#endif
+
+/* External wakeup pin edge on intent found.  0 for rising edge, 1 for falling edge */
+#ifndef appconfINTENT_WAKEUP_EDGE_TYPE
+#define appconfINTENT_WAKEUP_EDGE_TYPE     0
+#endif
+
+/* Delay between external wakeup pin edge and intent output */
+#ifndef appconfINTENT_TRANSPORT_DELAY_MS
+#define appconfINTENT_TRANSPORT_DELAY_MS     50
+#endif
+
+#ifndef appconfINTENT_I2C_OUTPUT_ENABLED
+#define appconfINTENT_I2C_OUTPUT_ENABLED   1
+#endif
+
+#ifndef appconfINTENT_I2C_OUTPUT_DEVICE_ADDR
+#define appconfINTENT_I2C_OUTPUT_DEVICE_ADDR 0x01
+#endif
+
+#ifndef appconfINTENT_UART_OUTPUT_ENABLED
+#define appconfINTENT_UART_OUTPUT_ENABLED   1
+#endif
+
+#ifndef appconfUART_BAUD_RATE
+#define appconfUART_BAUD_RATE       9600
+#endif
 
 /**
  * A positive delay will delay mics
@@ -68,10 +128,6 @@
 
 #ifndef appconfUSB_ENABLED
 #define appconfUSB_ENABLED         0
-#endif
-
-#ifndef appconfWW_ENABLED
-#define appconfWW_ENABLED          1
 #endif
 
 #ifndef appconfUSB_AUDIO_SAMPLE_RATE
@@ -133,7 +189,7 @@
 #include "app_conf_check.h"
 
 /* WW Config */
-#define appconfWW_FRAMES_PER_INFERENCE          (160)
+#define appconfWW_FRAMES_PER_INFERENCE          (240)
 
 /* I/O and interrupt cores for Tile 0 */
 /* Note, USB and SPI are mutually exclusive */
@@ -161,6 +217,7 @@
 #define appconfUSB_AUDIO_TASK_PRIORITY            (configMAX_PRIORITIES/2 + 1)
 #define appconfSPI_TASK_PRIORITY                  (configMAX_PRIORITIES/2 + 1)
 #define appconfQSPI_FLASH_TASK_PRIORITY           (configMAX_PRIORITIES/2 + 0)
-#define appconfWW_TASK_PRIORITY                   (configMAX_PRIORITIES/2 - 1)
+#define appconfINTENT_MODEL_RUNNER_TASK_PRIORITY  (configMAX_PRIORITIES - 2)
+#define appconfLED_TASK_PRIORITY                  (configMAX_PRIORITIES / 2 - 1)
 
 #endif /* APP_CONF_H_ */