+++ test/ffd_gpio/gpio.cmake
@@ -2,7 +2,7 @@
 # Gather Sources
 #**********************
 file(GLOB_RECURSE APP_SOURCES ${CMAKE_CURRENT_LIST_DIR}/src/*.c)
-list(APPEND APP_SOURCES ${SOLUTION_VOICE_ROOT_PATH}/examples/ffd/src/intent_handler/intent_handler.c)
+list(APPEND APP_SOURCES ${SOLUTION_VOICE_ROOT_PATH}/modules/asr/intent_handler/intent_handler.c)
 set(APP_INCLUDES ${CMAKE_CURRENT_LIST_DIR}/src
                  ${CMAKE_CURRENT_LIST_DIR}/src/dummy_headers
                  ${SOLUTION_VOICE_ROOT_PATH}/modules/asr
@@ -64,6 +64,7 @@ target_sources(${TARGET_NAME} PUBLIC ${APP_SOURCES})
 target_include_directories(${TARGET_NAME} PUBLIC ${APP_INCLUDES})
 target_compile_definitions(${TARGET_NAME} PUBLIC ${APP_COMPILE_DEFINITIONS} THIS_XCORE_TILE=0)
 target_compile_options(${TARGET_NAME} PRIVATE ${APP_COMPILER_FLAGS})
+
 target_link_libraries(${TARGET_NAME} PUBLIC core::general rtos::freertos rtos::drivers::audio)
 target_link_options(${TARGET_NAME} PRIVATE ${APP_LINK_OPTIONS})
 unset(TARGET_NAME)